+++ server/src/test/java/org/opensearch/discovery/DiscoveryModuleTests.java
@@ -34,6 +34,7 @@ package org.opensearch.discovery;
 import org.opensearch.Version;
 import org.opensearch.cluster.ClusterState;
 import org.opensearch.cluster.coordination.Coordinator;
+import org.opensearch.cluster.coordination.PersistedStateRegistry;
 import org.opensearch.cluster.node.DiscoveryNode;
 import org.opensearch.cluster.routing.RerouteService;
 import org.opensearch.cluster.service.ClusterApplier;
@@ -120,7 +121,8 @@ public class DiscoveryModuleTests extends OpenSearchTestCase {
             createTempDir().toAbsolutePath(),
             gatewayMetaState,
             mock(RerouteService.class),
-            null
+            null,
+            new PersistedStateRegistry()
         );
     }